backend/fixieAvatar/requirements.txt
djangorestframework
djangorestframework-simplejwt
mssql-django
pyodbc
gunicorn==23.0.0
kafka-python
cryptography